iccsdfai2025/airport/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
pytesseract==0.3.10
opencv-python==4.7.0.72
Pillow==9.5.0